refactor: introduce utils_build and utils_fs for command setup and file handling
- Added utils_build.go to encapsulate command setup logic, including the new SetUpCustomCommands function that utilizes CommandDetails for defining custom commands.
- Introduced utils_fs.go for file and directory management, implementing functions for creating random directories and writing files, enhancing modularity and maintainability.
- Removed redundant directory handling and command setup logic from utils.go, streamlining the codebase and improving clarity.
- This refactoring aligns with previous efforts to standardize command setup across stages, enhancing consistency and future extensibility.
